Slight refactoring for ease-of-use.
Renamed results from entries list and checkbox list to 'values' (for uniformity). Default selection in the entries list may now be either an index -or- a text string. Entries list will now automatically use a scroll-bar if the number of items exceeds the height and the scrollbar argument is not explicitly set.